Move some routing logic into a new sopro.js module
